fix: Role counts not updating when updating tags
Role counts weren't updating when using the Tag Manager to
update tags for a renamed role.

Tag Manager was updating role vm names by setting the public
Name property, which was firing a change event that the Deck
Builder view model was handling. The Deck Builder VM only
expected this event to occur when the Applies value updated,
so it was modifying the role count when it shouldn't.

Rename role VM without raising property change event.