feat: Only apply Scryfall tags when no role exists
Updating roles with Scryfall would never uncheck a role that
has already been checked, but it would always check a role
even if it had been manually unchecked. This change identifies
what source each role was set by, and if the source is either
the user or the role database, Scryfall tags won't overwrite it.

This can be overridden using the new checkbox in the Manage Tags
window, but includes a warning dialog, because it will clobber
all role data for the deck with Scryfall data.